Abstract fund recovery test
